+++ crates/client/rpc/src/utils/blockifier_state_adapter.rs
@@ -1,7 +1,7 @@
 use std::collections::HashSet;
 use std::sync::Arc;
 
-use blockifier::execution::contract_class::ContractClass;
+use blockifier::execution::contract_class::{ContractClass, ContractClassV0, ContractClassV1};
 use blockifier::state::cached_state::CommitmentStateDiff;
 use blockifier::state::errors::StateError;
 use blockifier::state::state_api::{State, StateReader, StateResult};
@@ -122,7 +122,22 @@ impl StateReader for BlockifierStateAdapter {
         match self.contract_class_update.get(&class_hash) {
             Some(contract_class) => Ok(contract_class.clone()),
             None => match self.backend.contract_class_data().get(&class_hash) {
-                Ok(Some(contract_class_data)) => Ok(contract_class_data.contract_class),
+                Ok(Some(contract_class_data)) => {
+                    let contract_class = if contract_class_data.sierra_program_length > 0 {
+                        ContractClass::V1(
+                            ContractClassV1::try_from_json_string(&contract_class_data.contract_class).map_err(
+                                |_| StateError::StateReadError("Failed to convert contract class V1".to_string()),
+                            )?,
+                        )
+                    } else {
+                        ContractClass::V0(
+                            ContractClassV0::try_from_json_string(&contract_class_data.contract_class).map_err(
+                                |_| StateError::StateReadError("Failed to convert contract class V0".to_string()),
+                            )?,
+                        )
+                    };
+                    Ok(contract_class)
+                }
                 _ => Err(StateError::UndeclaredClassHash(class_hash)),
             },
         }