+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Next release
 
+- feat(class): change class definition storage
 - fix: pending block sync
 - fix: transaction traces
 - feat: store tx receipts